Update mongorestore with all major changes to mongoimport (#11)
## Major changes
---
- Refactored most shared Cosmos DB logic into the `common/cosmosdb` folder
- Refactored all shared flags (`--autoScaleWorker`, `--throughput`, `--shardKey`, ...etc) into ToolOptions

### Minor changes
---
- Removed remnants of old retry code inside `mgo`
- Removed logger from `mgo/session`
- Moved function execution benchmark into common
